ModelArray: Statistical Analysis of Element-Wise Neuroimaging Data

The ModelArray package provides an S4 class and associated methods for performing massively univariate statistical analyses on element-wise (fixel, voxel, or vertex) neuroimaging data stored in HDF5 files.

Details

The core workflow is:

Inspect an HDF5 file with h5summary()

Load data with ModelArray()

Fit models with ModelArray.lm, ModelArray.gam, or ModelArray.wrap

Access results via scalars, results, sources

For multi-session or multi-modality analyses, combine ModelArrays with mergeModelArrays before fitting cross-scalar models.

Centralized imports

The following imports are consolidated here because they cannot be expressed as pkg::fun() qualified calls. All other add-on package functions are called with explicit namespacing throughout the package source code.

References

Zhao, C., et al. (2023). ModelArray–An R package for statistical analysis of fixel-wise data. NeuroImage, 271, 120037. doi:10.1016/j.neuroimage.2023.120037

See also

Author

Maintainer: Matthew Cieslak Matthew.Cieslak@pennmedicine.upenn.edu

Authors:

Chenying Zhao

Tinashe Tapera

Theodore Satterthwaite